packages/core/pyproject.toml
[project]
name = "slide2anki-core"
version = "0.1.0"
description = "Core pipeline for converting lecture slides to Anki flashcards"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-openai>=0.1.0",
    "pydantic>=2.7.1",
    "pillow>=10.3.0",
    "pdf2image>=1.17.0",
    "genanki>=0.13.1",
    "httpx>=0.27.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.4",
    "black>=24.4.2",
    "mypy>=1.10.0",
]
ollama = [
    "langchain-ollama>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["slide2anki_core"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
